minor changes for python 3.4 compatability

+++ tests/test_hypergraph_util.py
@@ -124,7 +124,7 @@ class HypergraphUtilTest(unittest.TestCase):
     expected = csr_matrix([
         [0, 0, 0],  # node 0 not listed
         [0, 0, 1]  # node 1 in edge 2
-    ])
+    ], dtype=np.float32)
     SparseArrayEquals(self, actual, expected)
 
   def test_ToCsrMatrix_multiple(self):
@@ -138,7 +138,7 @@ class HypergraphUtilTest(unittest.TestCase):
         [0, 0, 0],  # node 0 not listed
         [0, 1, 1],  # node 1 in edge 1 & 2
         [1, 0, 0]  # node 2 in edge 0
-    ])
+    ], dtype=np.float32)
     SparseArrayEquals(self, actual, expected)
 
   def test_ToCsrMatrix_empty(self):
